+++ src/App.vue
@@ -5,7 +5,7 @@
 </template>
 
 <script>
-import NavBar from "@/components/Navbar";
+import NavBar from "@/components/NavBar";
 
 export default {
   name: 'App',
