+++ src/components/NavBar.vue
